Worked on autoLogout functionality, added more documentation
Today we mostly just worked on autoLogout. We are currently running the
autoLogout main on another thread. Our initial test indicates that it
works (and we didn't completely break Django), but more tests are
necessary to make sure we didn't miss something. Also added more
comments to the models and views files.