CSS Testing
Now I'm just trying to see if my stylesheet is working at all (if it is,
I'm going to revert this commit, as I am currently changing the color to
fuschia to test it). Shelving the changing background-color on
login/logout issue for now, I'm tired. I'll post it as an issue.