Added color change on successful login/logout
Added a color argument to the return render statements. I have run into a problem with further features I have in mind, and I think I may need to implement the logInPage as a class view, as opposed to a function view (which is what it is currently). Would make things a lot easier and a lot nicer.